requirements.txt
sentence-transformers>=3.0.0
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
umap-learn>=0.5.0
matplotlib>=3.7.0
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
